https://leetcode.com/problems/number-of-islands/
Problem: Number of Islands
The goal of this problem is to count the number of islands in a 2D grid. An island is a group of '1's (representing land) that are adjacent horizontally or vertically.

Approach: Depth-First Search (DFS)
DFS is a common algorithm for traversing or searching tree or graph data structures. In this problem, we use a depth-first search to traverse the grid and mark visited cells as part of an island.