https://leetcode.com/problems/jump-game/
55. Jump Game
Problem Description:
Given an array of non-negative integers nums, where each element at index i represents the maximum number of steps you can jump from that index, determine if you can reach the last index.
Input: nums = [2,3,1,1,4]
Output: true
Explanation: Jump 1 step from index 0 to 1, then 3 steps to the last index.
Algorithm (Optimized Greedy Approach):
The provided algorithm is a Greedy approach to solve the problem.

We initialize a variable goalPost to be the last index of the array, indicating the position we need to reach.

Starting from the second-to-last element of the array, we iterate backward to the first element.

For each element at index i, we check if it's possible to reach the goalPost from that index:

If i + nums[i] (the maximum jump from index i) is greater than or equal to goalPost, it means we can reach the goalPost from index i. So, we update goalPost to be i.
If at the end goalPost is 0, it means we can reach the last index from the beginning of the array.