https://leetcode.com/problems/valid-palindrome/
125. Valid Palindrome
Problem Statement:
Given a string, determine if it is a palindrome, considering only alphanumeric characters and ignoring cases.

Example:
Input:
s = "A man, a plan, a canal: Panama"
Output: true
Algorithm and Approach:
The provided algorithm is a simple approach to determine whether the input string is a valid palindrome.

We start by removing all non-alphanumeric characters from the input string using a regular expression (/[^A-Z0-9]/ig). This gives us a new string with only alphanumeric characters.

We convert the new string to lowercase to ignore cases.

We reverse the new string and check if it is equal to the original new string.

If the reversed new string is equal to the original new string, the input string is a valid palindrome, and we return true. Otherwise, we return false.